Split boardpreviewer out of boardinit and variantpreviewer and of variantreader
