+++ src/client/scripts/esm/game/chess/engines/engineCheckmatePractice.ts
@@ -1669,16 +1669,12 @@ function runIterativeDeepening(
 							);
 						}
 					}
-					const emptyPieceMovesets = {}; // <--- Is this gonna be an issue?
 					const basegame = input_gamefile.basegame;
 					const dummy_board = {
 						moves: [],
-						// pieceMovesets is the only required gamefile property that is lost when sending the gamefile to the engine.
-						// This will cause the possible slides to be calculated incorrectly, and thus the `lines` property not entirely filled out.
-						// I THINK we are safe though, because I saw nowhere in detectInsufficientMaterial() where it reads the lines.
+						// Slide lines are not needed here — detectInsufficientMaterial() never reads lines.
 						pieces: organizedpieces.processInitialPosition(
 							piecesOrganizedByKey,
-							emptyPieceMovesets,
 							basegame.gameRules.turnOrder,
 							input_gamefile.boardsim.editor,
 							basegame.gameRules.promotion,
