+++ src/client/scripts/esm/game/chess/pastegame.ts
@@ -13,9 +13,9 @@ import type { Additional, VariantOptions } from '../../../../../shared/chess/log
 
 import boardutil from '../../../../../shared/chess/util/boardutil.js';
 import variantcache from '../../../../../shared/chess/variants/variantcache.js';
-import variantreader from '../../../../../shared/chess/variants/variantreader.js';
 import variantregistry from '../../../../../shared/chess/variants/variantregistry.js';
-import { pieceCountToDisableCheckmate } from '../../../../../shared/chess/logic/checkmate.js';
+import variantpreviewer from '../../../../../shared/chess/variants/variantpreviewer.js';
+import { pieceCountToDisableCheckmate } from '../../../../../shared/chess/util/winconutil.js';
 import icnconverter, {
 	MoveParsed,
 	LongFormatOut,
@@ -219,7 +219,7 @@ async function getPositionAndSpecialRightsFromLongFormat(
 		// Before pasting a game, we must ensure the variant has been loaded
 		await variantcache.ensureVariantLoaded(variantCode);
 		const mod = variantcache.getModule(variantCode);
-		return variantreader.getStartingPositionOfVariant(mod, timestamp);
+		return variantpreviewer.getStartingPositionOfVariant(mod, timestamp);
 	} else {
 		// Empty position
 		return { position: new Map(), specialRights: new Set() };
