+++ src/client/scripts/esm/game/rendering/piecemodels.ts
@@ -5,9 +5,9 @@
  */
 
 import type { Piece } from '../../../../../shared/chess/util/boardutil.js';
-import type { Board } from '../../../../../shared/chess/logic/boardinit.js';
 import type { Coords } from '../../../../../shared/chess/util/coordutil.js';
 import type { TypeGroup } from '../../../../../shared/chess/util/typeutil.js';
+import type { BoardPreview } from '../../../../../shared/chess/logic/boardpreviewer.js';
 
 import vectors from '../../../../../shared/util/math/vectors.js';
 import typeutil from '../../../../../shared/chess/util/typeutil.js';
@@ -104,7 +104,7 @@ const ATTRIBUTE_INFO: AttributeInfoInstanced = {
  *
  * SLOWEST. Minimize calling.
  */
-function regenAll(boardsim: Board, mesh: Mesh | undefined): void {
+function regenAll(boardsim: BoardPreview, mesh: Mesh | undefined): void {
 	if (!mesh) return;
 	console.log('Regenerating all piece type meshes.');
 
@@ -135,7 +135,7 @@ function regenAll(boardsim: Board, mesh: Mesh | undefined): void {
  * @param mesh
  * @param type - The type of piece to regen the model for (e.g. 'pawnsW')
  */
-function regenType(boardsim: Board, mesh: Mesh, type: number): void {
+function regenType(boardsim: BoardPreview, mesh: Mesh, type: number): void {
 	console.log(`Regenerating mesh of type ${type}.`);
 
 	if (typeutil.getRawType(type) === r.VOID)
@@ -154,7 +154,7 @@ function regenType(boardsim: Board, mesh: Mesh, type: number): void {
  * @param mesh
  * @param type - The type of piece of which to generate the model for (e.g. "pawnsW")
  */
-function genTypeModel(boardsim: Board, mesh: Mesh, type: number): MeshData {
+function genTypeModel(boardsim: BoardPreview, mesh: Mesh, type: number): MeshData {
 	const vertexData = instancedshapes.getDataTexture(mesh.inverted);
 	const instanceData: InstanceData = getInstanceDataForTypeRange(boardsim, mesh, type);
 
@@ -178,7 +178,7 @@ function genTypeModel(boardsim: Board, mesh: Mesh, type: number): MeshData {
  *
  * SLOWEST. Minimize calling.
  */
-function genVoidModel(boardsim: Board, mesh: Mesh, type: number): MeshData {
+function genVoidModel(boardsim: BoardPreview, mesh: Mesh, type: number): MeshData {
 	// const voidColor = preferences.getTintColorOfType(type); // Black, from the pieceTheme
 	const voidColor = gl.getParameter(gl.COLOR_CLEAR_VALUE); // Same color as the sky / void space star field. DOESN'T EVEN MATTER SINCE IT'S A MASK!
 	const vertexData: number[] = instancedshapes.getDataLegalMoveSquare(voidColor);
@@ -201,7 +201,11 @@ function genVoidModel(boardsim: Board, mesh: Mesh, type: number): MeshData {
  * The instance data contains only the offset of each piece instance, with a stride of 2.
  * Thus, this works will all types of pieces, even those without a texture, such as voids.
  */
-function getInstanceDataForTypeRange(boardsim: Board, mesh: Mesh, type: number): InstanceData {
+function getInstanceDataForTypeRange(
+	boardsim: BoardPreview,
+	mesh: Mesh,
+	type: number,
+): InstanceData {
 	// const range = boardsim.pieces.typeRanges.get(type)!;
 	// const instanceData64: Float64Array = new Float64Array((range.end - range.start) * STRIDE_PER_PIECE); // Initialize with all 0's
 	const instanceData: InstanceData = []; // Initialize empty
@@ -283,7 +287,7 @@ function castBigIntArrayToFloat32(instanceData: bigint[]): Float32Array {
  * uniform translations upon rendering, and reinits them on the gpu.
  * Faster than {@link regenAll}.
  */
-function shiftAll(boardsim: Board, mesh: Mesh): void {
+function shiftAll(boardsim: BoardPreview, mesh: Mesh): void {
 	console.log('Shifting all piece meshes.');
 
 	const newOffset = geometry.roundPointToNearestGridpoint(boardpos.getBoardPos(), REGEN_RANGE);
@@ -412,7 +416,7 @@ function deletebufferdata(mesh: Mesh, piece: Piece): void {
  * Renders ever piece type mesh of the game, EXCLUDING voids,
  * translating and scaling them into position.
  */
-function renderAll(boardsim: Board, mesh: Mesh | undefined): void {
+function renderAll(boardsim: BoardPreview, mesh: Mesh | undefined): void {
 	if (!mesh) return; // Mesh hasn't been generated yet
 
 	const { position, scale } = meshes.getBoardRenderTransform(mesh.offset, Z);
